AttryMe Validate
This is a function built to use custom attributes to validate a function, allowing one single js file to be used and referenced throughout an entire site. You control how it validates and what it validates via the attributes of every form element.

This has been tested and works on several sites. Support is IE9+, FF30+, Chrome 35+